+++ src/services/config.ts
@@ -14,6 +14,8 @@ const CONFIG_MAP = {
   YOUTUBE_API_KEY: process.env.YOUTUBE_API_KEY,
   SPOTIFY_CLIENT_ID: process.env.SPOTIFY_CLIENT_ID,
   SPOTIFY_CLIENT_SECRET: process.env.SPOTIFY_CLIENT_SECRET,
+  SOUNDCLOUD_CLIENT_ID: process.env.SOUNDCLOUD_CLIENT_ID,
+  SOUNDCLOUD_OAUTH_TOKEN: process.env.SOUNDCLOUD_OAUTH_TOKEN,
   WEBSERVER_PORT: process.env.WEBSERVER_PORT ?? 80,
   WEBSERVER_PASSWORD: process.env.WEBSERVER_PASSWORD ?? '',
   REGISTER_COMMANDS_ON_BOT: process.env.REGISTER_COMMANDS_ON_BOT === 'true',
@@ -45,6 +47,8 @@ export default class Config {
   readonly YOUTUBE_API_KEY!: string;
   readonly SPOTIFY_CLIENT_ID!: string;
   readonly SPOTIFY_CLIENT_SECRET!: string;
+  readonly SOUNDCLOUD_CLIENT_ID!: string;
+  readonly SOUNDCLOUD_OAUTH_TOKEN!: string;
   readonly REGISTER_COMMANDS_ON_BOT!: boolean;
   readonly DATA_DIR!: string;
   readonly CACHE_DIR!: string;